+++ src/app/(main)/cabinet/anagata/bod/page.tsx
