+++ src/app/(main)/puma/puma-id/page.tsx
